Refactored enum class into testUtils.py. GH #4973
